All done

+++ MsPrices/src/test/java/com/airtrips/prices/PricesApplicationTests.java
@@ -1,5 +1,6 @@
 package com.airtrips.prices;
 
+import com.airtrips.prices.services.PriceService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -18,8 +19,7 @@ class PricesApplicationTests {
 
 	@Test
 	void CalculatePrice_ReturnsNewValue_WhenUsingValidPrice() {
-		int price = service.calculatePrice(HAS_LUGGAGE, AGE, DISTANCE, DATE );
-		//
+		double price = service.calculatePrice(HAS_LUGGAGE, AGE, DISTANCE, DATE );
 		Assert.isTrue(price != 200, "The returned price is diferent than 200 € (base price)");
 	}
 }
